airflow-main/Dashboards/requirements.txt
pandas
structlog
streamlit
altair
numpy
mysql.connector